i2c: exynos-i2c: Fix race in clearing IRQ/signaling completion
The exynos5_i2c_irq() had a race in it where it would signal that it
was complete() and _then_ clear its IRQ.  This is not such a good idea
since the completion will effectively release control over the device.

I could confirm that this was a problem by adding a simple udelay(10)
in the driver after the complete() call.  The device wouldn't even
boot all the way.

To fix this, change the order of the complete call to avoid the race.

BUG=chrome-os-partner:22834,chrome-os-partner:22837
TEST=Try to find a place to put a delay to causes a boot crash; can't.
TEST=Stress the i2c bus by reading a lot.

Change-Id: I2b2cb1eaf5f77f14e9b8532e2fee86b31c5fa2b5
Signed-off-by: Doug Anderson <dianders@chromium.org>
Previous-Reviewed-on: https://chromium-review.googlesource.com/170307
(cherry picked from commit 8452d0df44ad4d3e28fc321cfde2b5109f404fa6)
Reviewed-on: https://chromium-review.googlesource.com/170471
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>