remount - don't error on xattr and acl options
Hi,

The current f2fs code errors if the xattr or acl options are passed when
remounting.  This is important in a typical scenario where f2fs is mounted
as a "ro" root file-system by the boot loader and then the init process wants
to remount it "rw" with the "remount,rw" option.

The attached patch fixes the problem.

Included here for review.

Signed-off-by: Jaegeuk Kim <jaegeuk.kim@samsung.com>