i2c: exynos5: Fix handling of TRANS_DONE and interrupts in general
We found that we would occasionally fail to see TRANS_DONE being set
while polling for it in exynos5_i2c_wait_bus_idle(), leading to the
theory that reading the TRANS_STATUS register at exactly the wrong
time could cause us to miss TRANS_DONE.  Instead of polling for it,
let's wait for the INT_I2C interrupt indicating the completion of
the transaction to fire.  This also requires interrupts to be enabled
before starting the transaction to ensure that we don't miss any
INT_I2C interrupts.  Because FIFO interrupts may occur before the
transaction is actually started, a spinlock is introduced which makes
enabling interrupts and starting the transfer atomic with respect to
the interrupt handler.

BUG=chrome-os-partner:22837
TEST=Insert WARNs into failure paths and see that nothing gets printed
out when stressing the i2c bus with reads and writes.

Change-Id: I2feb9ba2f99938df1db6b60ea49a7a05848a3fc6
Signed-off-by: Doug Anderson <dianders@chromium.org>
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Previous-Reviewed-on: https://chromium-review.googlesource.com/170556
(cherry picked from commit f890c7f952a99f2b0753f42f16b774d68f5ad166)
Reviewed-on: https://chromium-review.googlesource.com/170949