drm/i915: Restore global mappings upon resume
commit c6916417c8575637659686ff0e4f744babf0cb4e
Author: Ben Widawsky <ben@bwidawsk.net>
Date:   Tue Sep 24 09:58:00 2013 -0700

    drm/i915: Use the new vm [un]bind functions

changes the interpretation of the bind routines, and in particular we
only bind into the global GTT if we pass a flag to the routine. So upon
resume, we need to ask for whatever we thing is bound into the global
GTT (cursors, fences, scanouts) to be rebound or else glorious
corruption ensues.

v2.1: Clear the is-bound flags before attempting to rebind.

Fixes: i-g-t/gem_suspend
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=69834
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Ben Widawsky <ben@bwidawsk.net>
[danvet: As discussed on irc drop the clearing of the
has_aliasing_ppgtt_mapping, ppgtt ptes will survive s/r or
hibernation, and higher leves (pdes) get restore in the ->enable
callback.]
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>