ARM: dts: Update HDMI FET for rev6
The HDMI FET moved from FET2 to FET7 in rev6.  Change the device tree
to handle this.  Note:

* In older revs FET7 was unused and called "P5.0V_VIDEO_MID_ADD".
* In newer revs FET2 is unused.

We already had FET2 marked as "always-on", so make things simple and
mark FET2 and FET7 as "always-on".  Be nice and refer to the correct
FET in the hdmi node as well, though it's not super important (since
the FET is currently marked to be always on).

BUG=chrome-os-partner:22944
TEST=Test HDMI on rev5; test HDMI on rev6.

Change-Id: I6305d7b03c521e19e8ef25de109cc27512b99f52
Signed-off-by: Doug Anderson <dianders@chromium.org>
Previous-Reviewed-on: https://chromium-review.googlesource.com/170845
(cherry picked from commit a69239072145903a5d255767ad0ba7bbb0b4bfaa)
Reviewed-on: https://chromium-review.googlesource.com/170945
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>