ipc/sem.c: synchronize the proc interface
The proc interface is not aware of sem_lock(), it calls ipc_lock_object()
directly.  This means that simple semop() operations can run in parallel
with the proc interface.  Right now, this is uncritical, because the
implementation doesn't do anything that requires a proper synchronization.

But it is dangerous and therefore should be fixed.

Signed-off-by: Manfred Spraul <manfred@colorfullife.com>
Cc: Davidlohr Bueso <davidlohr.bueso@hp.com>
Cc: Mike Galbraith <efault@gmx.de>
Cc: Rik van Riel <riel@redhat.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>