qseecom: Copy userspace buffer into kernel space before dereferencing
ION memory is used for user space to kernel space data passing.
This is directly accessible in kernel. But, if the IOCTL is called
from user space without using User space library, then data might
be pointing to some other memory location, in which case, it would
not be possible to dereference this location in kernel & hence it
would be accessing invalid memory.

Change-Id: Ie02382f03ccfe0e2eaca54021ca67e765faf90bf
Signed-off-by: Hariprasad Dhalinarasimha <hnamgund@codeaurora.org>