usb: msm_otg: Remove sleep waiting loop from the ci13xxx_wakeup function
When USB core exits from low power state in BAM to BAM mode, the
ci13xxx_wakeup function gets called from an atomic context and therefore
the usleep() call fails and an error message is returned. The fix is to
remove the usleep() call when the CI13XXX driver is used.

Change-Id: I3624180622380616453e9725fe0aec82ff9af381
Signed-off-by: Danny Segal <dsegal@codeaurora.org>