CHROMIUM: mmc: fix sdio bus timeout by not setting card read threshold register
The previously merged tuning code is incorrectly setting the Card Read
Threshold register for all cases, when it should only do it for HS200 and
SDR104 and this patch removes that code.  This is only intended to support
SDR50 for wifi, which doesn't require card read threshold.

BUG=chrome-os-partner:20212
TEST=`cat /sys/kernel/debug/mmc1/ios |grep SDR50` has output

Signed-off-by: Sonny Rao <sonnyrao@chromium.org>

Change-Id: I75acb4867f0e5017dbdf9f2b62058bd1fb81036b
Reviewed-on: https://chromium-review.googlesource.com/172732
Reviewed-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Queue: Sonny Rao <sonnyrao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/173216
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>