UPSTREAM: cpufreq: tegra: fix the wrong clock name
The "cpu" and "pclk_p_cclk" was a virtual clock name that was used in
the legacy Tegra clock framework. It was not used after converting to
CCF. Fix it as the correct clock name that we are using.

Tested-by: Stephen Warren <swarren@nvidia.com>
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
(cherry picked from commit b192b910f3832793082c1a18262c4da0efe121ae)
Signed-off-by: Bill Huang <bilhuang@nvidia.com>

BUG=None
TEST=Boot up Tegra20 to see if cpufreq driver is working correctly

Change-Id: I9add2e00de19296217089d18230f1133e188fccd
Reviewed-on: https://chromium-review.googlesource.com/174300
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>