dm cache: add trace policy shim
This commit includes a non-terminal policy (aka "shim") called trace
that may be stacked ontop of a terminal policy (e.g. mq).

The second shim, trace, adds function-call-level tracing to the policy
stack.  By default, an INFO-level log message including function name
and parameter(s) will be generated whenever most of the cache policy
interface functions are called.  An interface to increase or decrease
the verbosity of the trace output is also provided.

trace_printk is used to provide the trace messages as it is more
scalable than printk.  This trace policy shim is intended for use during
the development and debugging of new cache policies.  Therefore, use of
trace_printk within a separate and temporarily used policy shim reduces
the traditional concerns associated with leaving trace_printk calls
dangling in production code.

Signed-off-by: Morgan Mears <morgan.mears@netapp.com>
Signed-off-by: Mike Snitzer <snitzer@redhat.com>