cpufreq: exynos-iks: use cluster switching
Due to ARM erratum 814169, running in a mixed A15/A7 configuration
can lead to a hardware deadlock.  In order to minimize the chance
of hitting this erratum, switch to using 4+0 or 0+4 configurations
only.  Since this synchronization is done in the kernel, there is
still a brief period where we could hit this erratum, but it is
believed to be extremely rare.

Unsetting "use_cluster_switching" (which could be determined at
runtime based on hardware revision in the future) will revert to
the old (IKS) behavior.

BUG=chrome-os-partner:24157
TEST=Boot on Pit, change governor to userspace and observe that
setting the frequency of one CPU changes all of them. Run s2r and
mem_stress_switcher to verify that they still work.

Change-Id: I9abe23afb70f17417c16c8a2b7e21862c534cabc
Signed-off-by: Arjun.K.V <arjun.kv@samsung.com>
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Signed-off-by: Akshu Agrawal <akshu.a@samsung.com>
Reviewed-on: https://chromium-review.googlesource.com/177330
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
(cherry picked from commit 11c9efae4f3dbd3072a0c4639151e8c6fe8aa0aa)
Reviewed-on: https://chromium-review.googlesource.com/178423