tegra124: Fix the table used for setting the frequency of PLLP.
This table had settings for 216MHz, but PLLP is (and is supposed to be)
configured 408MHz. When that table is used, the kernel can't find appropriate
settings and panics.

BUG=None
TEST=Before this change and when booting using coreboot, the settings of the
clocks caused the kernel to attempt to use this table and would panic. After
this change it no longer panics.

Change-Id: Ia9c8b0408c0c06957c9ba4bc07d2ad9894e7f594
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/174829
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>