ARM: dts: tegra124-venice2: remove eMMC regulators
VCC for eMMC is not 1.8V_VDDIO, but instead 3.3V_RUN - 1.8V_VDDIO is
actually VCCQ for eMMC.  Remove vmmc-supply for now since 3.3V_RUN is
not a regulator.  We can later add it as a GPIO-controlled fixed
regulator if necessary.

Also remove sd-slot-supply as it is not clear what its purpose is.

BUG=None
TEST=Boot Nyan from SD card, see that eMMC is probed as mmcblk1.
Installing to and booting from eMMC also works.

Change-Id: Ia7c26a631562c0092ad77eba9d8e6c382cf8c206
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/174846
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>