ARM: dts: msm: Split APQ8084 CDP and MTP DT files based on SMB chips used.
The CDP and MTP boards need to be further split based on whether
they have SMB1357 or SMB349 charger chips.

For a CDP
The platform subtype value of 0 and 1 indicates presence of SMB349
while 2 and 4 indicates presence of SMB1357.

For a MTP
The platform subtype value 1 indicates presence of SMB349
while 2 indicates presence of SMB1357.

While at it, add SMB1357 device definitions for boards that have it
and move the already existing SMB349 device definitions to the
appropriate files. These devices are used by USB driver for its
otg supply, move the phandles referencing them accordingly.

Change-Id: I5485c94b376bed2ff9514c916b0f76e39a86b719
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
Signed-off-by: Piyush Gupta <guptapi@codeaurora.org>