CHROMIUM: i2c-designware: make *CNT values configurable
Calculating *CNT values based solely on the source clock (as it is done in
the driver now) is not always the most accurate method. One should also
take the bus pull-up resistor values, capacitance etc. into account. These
properties are very platform specific.

With this patch we let the platform code to specify more accurate *CNT
values if they are known beforehand, otherwise we stick with the original
method and calculate the values based on the source clock.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>

BUG=
TEST=

Change-Id: I3996dddec769deddad348e5a030a8372972b4e46
Reviewed-on: https://chromium-review.googlesource.com/180695
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Benson Leung <bleung@chromium.org>