block: fix double increment of bio vec count on bouncing
Tushar/Tomasz both reported a boot failure on an Arndale
board on mmcblock. Turns out the bounce code got broken
with commit febca1baea1c that copied the vec count
from the bio_src to to bio_dest on a bio clone, since
the new clone logic increments vec count as well. This
resulted in a double increment, screwing up DMA mapping
of the bio.

Reported-by: Tushar Behera <tushar.behera@linaro.org>
Reported-by: Tomasz Figa <t.figa@samsung.com>
Tested-by: Stephen Warren <swarren@nvidia.com>
Signed-off-by: Jens Axboe <axboe@kernel.dk>