UPSTREAM: ftrace: Add private data to ftrace_ops
Passing data to the function callback was originally done by adding the
ftrace_ops in another structure, and using the container_of() to get
the field. But this adds a bit more complexity than it is worth, and
adding a simple .private field to ftrace_ops makes things a lot easier.

But be warned, the .private data should not be freed once it is used
unless the ftrace_ops itself has gone through the necessary freeing
routines. A simple synchronize_sched() is not enough as functions
can be traced that are called outside the view of RCU and all its
concoctions.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>
(cherry picked from commit b7e00a6c53e9134d5cf7631582acaf027a5ded26)

Change-Id: I9364467876021e45e74653a83e7e0b44d07e7c72
Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Signed-off-by: David Riley <davidriley@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/211343
Reviewed-by: Olof Johansson <olofj@chromium.org>