msm: mdss: Support cursor hotspot feature
Handle FB_CUR_SETHOT flag to enable cursor hotspot support.

The cursor hotspot is a point on the cursor image used as the active
aiming pixel for user input focused on the display.

The default cursor hotspot is at (0, 0) i.e., the upper left corner
of the cursor image. This works for a typical cursor image like a
mouse pointer with the tip of the pointer arrow at the (0, 0)
position of the cursor image. If the cursor is a cross-hair, its
hotspot would be at (31, 31) i.e., the centre of the 64x64 cursor
image. For a 64x64 cursor, the hotspot is specified from (0, 0) to
(63, 63). The cursor's hotspot pixel is displayed at the
(img->dx, img-dy) cursor position on the screen.

Change-Id: I366d3e3faea17c4fa083adaec30e9ce83f742d6f
Signed-off-by: Beeresh Gopal <gbeeresh@codeaurora.org>
Signed-off-by: Mathew Karimpanal <kjmathew@codeaurora.org>