MALI: Separate PM callbacks for suspend/resume
The existing platform-specific API exposed callbacks to turn the GPU
on and off. These were used by the power-policy (on idle), and on system
suspend.

However, it is likely a platform needs to perform different operations
on suspend. In the case of ChromiumOS on Exynos we need to reset the
DVFS before suspending. We also need to ensure we always reset the GPU
on resume, as we don't reset the GPU after idle (as the GPU is not
powered off).

BUG=chromium:316186
BUG=chrome-os-partner:23950
TEST=build and run on daisy; ensure the system suspends correctly

Signed-off-by: David Garbett <david.garbett@arm.com>
Previous-Reviewed-on: https://chromium-review.googlesource.com/176830
(cherry picked from commit 28edd5b2e5ef6c386375f19f62c561eaf06cde3c)

Change-Id: Ib6370f141dbed42e52d3de35aafe57d85137c83e
Reviewed-on: https://chromium-review.googlesource.com/177980
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>
Commit-Queue: Doug Anderson <dianders@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>