CHROMIUM: LPE: Eliminate non-portable virt_to_phys in sst_pcm_new
The standard model for a <xxx>_pcm_new procedure is to use
dma-coherent allocation, and then use the resulting virtual
address for CPU access to the memory and the (suitable cast)
resulting dma addr for the physical address.  This patch
changes the sst_pcm_new procedure and subsequent use of
allocated memory to use that standard model instead of using
non-portable virt_to_phys to translate CPU-access pointer to
physical-address reference for the DSP.

This change leaves two remaining virt_to_phys() references
in sound/soc/mid-x86/ directory structure.  These are used
in referencing a ring-buffer used to push data to a
compression module.  The object in question is a (struct
snd_compr_stream *)->runtime->buffer, which is allocated by
sound/soc core procedure using kmalloc() with no option to
use dma-coherent allocation.

It would take an extensive implementation around the
snd_compr_ops.copy() operation to fix this.  The impetus to
fix this would be the debate whether the kmalloc() followed
by virt_to_phys() usage is portable for all future time
across all platforms.  It doesn't look like this will be a
problem for the forseeable future on x86 platforms, and the
code in question has other x86-specific characteristics, so
it did not seem worthwhile to tackle these remaining
virt_to_phys() in this patch.

BUG=chrome-os-partner:24118
TEST=aplay -l
     # lists 'baytrailaudio'

Change-Id: Id9b343b6ce7aee7386be9bf00a1813c7f8bc153c
Reviewed-on: https://chromium-review.googlesource.com/180717
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>