gpiolib: fix of_find_gpio() when OF not defined
The prototype for static GPIO lookup functions has been updated to use
an explicit type for GPIO lookup flags. Unfortunately the definition of
of_find_gpio() when CONFIG_OF is not defined has been omitted, which
triggers a warning. This patch fixes this.

Signed-off-by: Alexandre Courbot <acourbot@nvidia.com>
--
This fixes the kbuild test robot warning about
"gpiolib: use dedicated flags for GPIO properties". Hopefully this can
be squashed into this same patch. Apologies for the trouble.
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>