msm-camera: Set Recommended value for t_CLK_MISS_COUNT
Receiver will disable the Rx-term after t_CLK_MISS_COUNT
(in MIPI_CSIPHY_LNCK_CFG4) expires. The default for this is set
to 40ns in hardware, but this should be over written in SW to maximize
the timing here. When this happens the Rx-term will disable during
the LP00 ->LP11 transition and t_CLK_MISS_COUNT won't affect anything.

Change-Id: I209cc06eaf60521d4cd4fa218f93f89ee9e1eae6
Signed-off-by: Chandan Gera <cgera@codeaurora.org>