CHROMIUM: bq24735: don't attempt I2C accesses if charger is not present
If we have an AC-detect GPIO for the charger and the charger is not
present during probe time, don't attempt I2C accesses to the charger
as they are not guaranteed to succeed.  The charger will be configured
once it is plugged in.  Also initialize the AC-detect GPIO as an input
so that its value can be read reliably before setting it up as an IRQ.

BUG=None
TEST=Boot on Nyan rev1 with AC unplugged and observe that the charger
and battery still probe.  Plug in the charger and see that charging
starts.

Change-Id: I350d6529d13bd831fb0b21b15c0e98ad615e3d0b
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/177822
Reviewed-by: Rhyland Klein <rklein@nvidia.com>
Reviewed-by: Dylan Reid <dgreid@chromium.org>