ALSA: hda - Add zenbook-boost model for ASUS Zenbook
There's been a complaint about the too soft output on ASUS Zenbook
with ALC269VB in comparison with Windows.  It turned out, however,
that the actual hardware output level is same between Linux and
Windows unless software plugin is used.  Still, it'd be convenient to
give a capability for increasing the output level via Realtek codec's
D-class amp.

As a compromise, here provides a new model option, zenbook-boost, to
enable the COEFs for the output boost amp.  It's not used unless the
option is given explicitly just for safety reason.

Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=62651
Signed-off-by: Takashi Iwai <tiwai@suse.de>