CHROMIUM: i2c: tegra: moving suspend/resume function from system to noirq
The waking up of I2C device would keep the interrupt when it woke up the
system. And trigger the interrupt after system finish _noirq resume
services. Then it caused I2C access error due to the I2C host not
resumed yet. Moving it from system to noirq to solve this issue.

BUG=chrome-os-partner:31211
TEST=resume from LP0 and check there is no error message to access irq
     of i2c devices when resuming from i2c devices

Change-Id: I8bbb6127cf72f5653ea94b991dd9ecbddbc6ca1a
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/221692
Reviewed-by: David Riley <davidriley@chromium.org>