CHROMIUM: i2c: tegra: moving suspend/resume function from system to noirq
The waking up of I2C device would keep the interrupt when it woke up the
system. And trigger the interrupt after system finish _noirq resume
services. Then it caused I2C access error due to the I2C host not
resumed yet. Moving it from system to noirq to solve this issue.

BUG=chrome-os-partner:40741
TEST=resume from LP0 and check there is no error message to access irq
     of i2c devices when resuming from i2c devices

Change-Id: If0d74c221221f9f05e96e494294a5f53632a0f4a
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/276942
Reviewed-by: Benson Leung <bleung@chromium.org>